Tweak YellowGreen line in color-names to simplify test
The YellowGreen line in the color-names test differed from all the
other name tests in that it doesn't perform a mathematical operation;
this means this line isn't neutral on whether hex codes should be left
as-is or converted to colour names if processed without any needed
changes (tested in other tests).  Adding the + 0 keeps the test neutral
by ensuring conversion always occurs.