Allow testing for exitcode and stderr output
If files "error" and "status" are existing
in the test directory, do expect errors
and test against them.

The errors returned by the Sass engines are
now treated as test failures (not fatal errors).

--unexpected-pass reports an error whenever
a test marked as "todo" does pass. Failing
"todo" errors are silently marked as "passed".

Also re-create output test files with --nuke.
"status" and "error" files are re-created if necessary.
Any failures will be reported in this case normally.

Fixes: https://github.com/sass/sass-spec/issues/136