--skip is now no-op
Convert --skip/-s to no-op, as this is now the default
behavior.

Rationale: As we now expect errors to be thrown by
the spec scripts it makes no sense to stop whenever
a first error is encountered.

Unfortunately we cannot print deprecation warning
in order not to break formats like tap or mess
with stderr (Windows PowerShell thinks for example
that if there is anything on stderr, the command failed).