Don't use the sass cache when running specs.
Some warnings are issued by the parser and those warnings aren't issued
with a warm cache because the parser is never used; this results in
spurious failures. Note that turning off the cache actually speeds up
sass-spec because there's very little code sharing via @import (not so
for most projects).