Enable the troublesom str-slice spec for libsass
This spec was updated with incorrect output in #997 due to a Ruby
Sass bug. In #1004 the output was updated to be correct and the spec
disabled for all implementations because they all suffered from the
original bug.

This PR enables it for LibSass which is now passing with correct output.