Update an extend spec for Dart Sass.
This is a particularly strange case. Dart Sass produces semantically
different output than Ruby Sass and LibSass, but it's because of a
non-semantic optimization: no implementation includes *all* selectors
that would be necessary in this case to thoroughly model the underlying
semantics. The specific nature of those optimizations aren't really part
of the Sass language, so it's more or less valid for a given
implementation to choose different selectors to include or
exclude—although it could cause incompatibilities and so should be
avoided if possible.

In this case, Dart Sass chooses different selectors because of its eager
extend implementation. I believe it would be quite difficult to get it
to emit the same selectors as Ruby Sass, since the (arbitrary) choice of
which selectors not to include is based on the whole resolved selector.
So I think it's fair to allow different output.