Proposal: Add map-fetch as a stricter implementation of map-get
When map-get is called with a key that doesn't exist, the function
returns null. When there is a simple typo, the null value may still end
up as valid CSS leading to a confusing rendered result that is missing
the expected styling. The developer must then debug the typo to fix it.

To help catch typos earlier, provide a stricter version of map-get that
fails on a missing key rather than returning null. The error should
provide enough information to quickly spot the typo.

The name "fetch" is inspired from Ruby's Hash#fetch method:
https://ruby-doc.org/core-3.1.2/Hash.html#method-i-fetch

[skip dart-sass]